Move logic to build API-agnostic objects (entity, access control) into postgres-core-builder
