+++ crates/postgres-subsystem/postgres-core-builder/src/lib.rs
@@ -1,5 +1,12 @@
 pub mod access_builder;
+pub mod access_utils;
+pub mod aggregate_type_builder;
 pub mod database_builder;
 pub mod naming;
 pub mod resolved_builder;
 pub mod resolved_type;
+pub mod shallow;
+pub mod system_builder;
+pub mod type_builder;
+
+pub use system_builder::SystemContextBuilding;
