+++ crates/postgres-subsystem/postgres-core-model/src/subsystem.rs
@@ -7,12 +7,37 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use core_plugin_interface::{error::ModelSerializationError, system_serializer::SystemSerializer};
+use core_plugin_interface::{
+    core_model::{
+        access::AccessPredicateExpression,
+        context_type::ContextType,
+        mapped_arena::{MappedArena, SerializableSlab},
+    },
+    error::ModelSerializationError,
+    system_serializer::SystemSerializer,
+};
 use exo_sql::Database;
 use serde::{Deserialize, Serialize};
 
-#[derive(Serialize, Deserialize, Debug, Default)]
+use crate::{
+    access::{DatabaseAccessPrimitiveExpression, InputAccessPrimitiveExpression},
+    aggregate::AggregateType,
+    types::{EntityType, PostgresPrimitiveType},
+};
+
+#[derive(Serialize, Deserialize, Debug)]
 pub struct PostgresCoreSubsystem {
+    pub contexts: MappedArena<ContextType>,
+    pub primitive_types: SerializableSlab<PostgresPrimitiveType>,
+    pub entity_types: SerializableSlab<EntityType>,
+
+    pub aggregate_types: SerializableSlab<AggregateType>,
+
+    pub input_access_expressions:
+        SerializableSlab<AccessPredicateExpression<InputAccessPrimitiveExpression>>,
+    pub database_access_expressions:
+        SerializableSlab<AccessPredicateExpression<DatabaseAccessPrimitiveExpression>>,
+
     pub database: Database,
 }
 
@@ -27,5 +52,21 @@ impl SystemSerializer for PostgresCoreSubsystem {
         reader: impl std::io::Read,
     ) -> Result<Self::Underlying, ModelSerializationError> {
         bincode::deserialize_from(reader).map_err(ModelSerializationError::Deserialize)
+    }
+}
+
+impl Default for PostgresCoreSubsystem {
+    fn default() -> Self {
+        Self {
+            contexts: MappedArena::default(),
+            primitive_types: SerializableSlab::new(),
+            entity_types: SerializableSlab::new(),
+            aggregate_types: SerializableSlab::new(),
+
+            input_access_expressions: SerializableSlab::new(),
+            database_access_expressions: SerializableSlab::new(),
+
+            database: Database::default(),
+        }
     }
 }
