+++ crates/postgres-subsystem/postgres-graphql-builder/src/lib.rs
@@ -9,8 +9,6 @@
 
 pub use plugin::PostgresGraphQLSubsystemBuilder;
 
-mod access_utils;
-mod aggregate_type_builder;
 mod builder;
 mod create_mutation_builder;
 mod delete_mutation_builder;
@@ -21,8 +19,9 @@ mod plugin;
 mod predicate_builder;
 mod query_builder;
 mod reference_input_type_builder;
-mod shallow;
 mod system_builder;
 mod type_builder;
 mod update_mutation_builder;
+
+mod shallow;
 mod utils;
