+++ crates/postgres-subsystem/postgres-graphql-resolver/src/abstract_operation_resolver.rs
@@ -30,7 +30,11 @@ pub async fn resolve_operation<'e>(
 
     let mut result = subsystem_resolver
         .executor
-        .execute(op, &mut tx, &subsystem_resolver.subsystem.database)
+        .execute(
+            op,
+            &mut tx,
+            &subsystem_resolver.subsystem.core_subsystem.database,
+        )
         .await
         .map_err(PostgresExecutionError::Postgres)?;
